Add LetsTrust TPM
LetsTrust-TPM is a TPM2.0 daughter board for the Raspberry Pi platform based on
an Infineon SLB9670 chip.
It is compatible with all Raspberry Pi models and probably the smallest
available addon board on the market.
The TPM can e.g. be used as a secure key storage and as a hwrng.

Signed-off-by: Peter Huewe <peterhuewe@gmx.de>