Update pin-8.md
The sentence ending in "…to achieve a stable." seemed off and prompted me to look into it. After a bit of reading, made a few changes, which i'm submitting here.

Also, "the UART pins are useful" --> "can be useful": setting up over SSH is also an option here, and i think it's better not to make it sound like UART pins are the main option.